requirements.txt
Flask==2.2.3
Flask-PyMongo==2.3.0
pandas==2.0.2
scikit-learn==1.2.2
numpy==1.24.3
PyPDF2==3.0.1
docx2txt==0.8
Werkzeug==2.2.3
pymongo==4.3.3
joblib==1.2.0
Flask-Mail
Flask-MongoEngine
tensorflow